requirements.txt
annotated-types==0.7.0
anyio==4.9.0
autopep8==2.3.2
bcrypt==4.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
Faker==37.4.0
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.3
httptools==0.6.4
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
motor==3.7.1
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
packaging==25.0
pycodestyle==2.13.0
pydantic==2.11.6
pydantic_core==2.33.2
pymongo==4.13.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tokenizers==0.21.1
torch==2.2.2
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
